Fix merge

+++ CoronavirusHerdImmunity/app/src/main/java/com/example/coronavirusherdimmunity/utils/BeaconDto.java
@@ -21,7 +21,7 @@ public class BeaconDto {
     public double x = 0;
     public double y = 0;
 
-    public BeaconDto(int identifier, int rssi, Distance distance, double distanceValue, double x, double y){
+    public BeaconDto(Long identifier, int rssi, Distance distance, double distanceValue, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = new Date().getTime() / 1000;
@@ -31,7 +31,7 @@ public class BeaconDto {
         this.y = y;
     }
 
-    public BeaconDto(int identifier, int rssi, long timestamp, Distance distance, double distanceValue, double x, double y){
+    public BeaconDto(Long identifier, int rssi, long timestamp, Distance distance, double distanceValue, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = timestamp;
@@ -42,7 +42,7 @@ public class BeaconDto {
 
     }
 
-    public BeaconDto(int identifier, int rssi, Date timestamp, Distance distance, double distanceValue, double x, double y){
+    public BeaconDto(Long identifier, int rssi, Date timestamp, Distance distance, double distanceValue, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = timestamp.getTime() / 1000;
